translated python code to javascript with a bit of refactoring
