+++ server/package.json
@@ -15,6 +15,7 @@
     "body-parser": "^1.20.2",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
+    "dataframe-js": "^1.4.4",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "gpt-3-encoder": "^1.1.4",
